:recycle: [Refactor] useTags 리팩터링 #main
