+++ src/entities/model/tag/atoms.ts
@@ -1,0 +1,4 @@
+import { atom } from "jotai"
+import type { Tag } from "../../../shared/types"
+
+export const tagsAtom = atom<Tag[]>([])
